REFACTOR #465 : 상품 / 리뷰 수 반환 API 리팩토링
