+++ src/main/java/com/lokoko/domain/product/domain/repository/ProductRepositoryCustom.java
@@ -31,4 +31,6 @@ public interface ProductRepositoryCustom {
 	Slice<ProductBrandInfoProjection> findProductsOrderedByRating(Pageable pageable);
 
 	Long countAllProducts();
+
+    int countProductsByBrandName(String brandName);
 }
